FastApiBackend/requirements.txt
fastapi>=0.104.1,<1.0.0
pydantic-settings>=2.1.0,<3.0.0
uvicorn>=0.15.0,<1.0.0
httpx>=0.23.0,<1.0.0
loguru
httpx==0.24.1
pydantic>=2.7.0
python-dotenv>=1.0.0,<2.0.0
paho-mqtt